Refactor: SplashScreen to check user login and survey status
This commit refactors the SplashScreen to determine navigation based on user login and survey completion status.
It introduces a new SplashViewModel with a SplashUiState to manage these conditions and interacts with FirebaseRepository for user status and UseCases for survey status.
Also, it refactors SplashScreen to observe SplashViewModel's SplashUiState and navigate the user to either Home or Survey screens accordingly.